comdist

comdist

ph_comdist(sample, phylo, null_model = 0, randomizations = 999,
  abundance = TRUE)

ph_comdistnt(sample, phylo, null_model = 0, randomizations = 999,
  abundance = TRUE)

Arguments

sample

(data.frame/character) sample data.frame or path to a sample file

phylo

(character/phylo) phylogeny as a phylo object or a newick string (will be written to a temp file if provided) - or a path to a file with a newick string

null_model

(integer) which null model to use. See Details.

randomizations

(numeric) number of randomizations. Default: 999

abundance

(logical) If TRUE (default) computed accounting for abundance. Otherwise, uses presence-absence.

Value

data.frame

Null models

0 - Phylogeny shuffle: This null model shuffles species labels across the entire phylogeny. This randomizes phylogenetic relationships among species.

1 - Species in each sample become random draws from sample pool: This null model maintains the species richness of each sample, but the identities of the species occurring in each sample are randomized. For each sample, species are drawn without replacement from the list of all species actually occurring in at least one sample. Thus, species in the phylogeny that are not actually observed to occur in a sample will not be included in the null communities

2 - Species in each sample become random draws from phylogeny pool: This null model maintains the species richness of each sample, but the identities of the species occurring in each sample are randomized. For each sample, species are drawn without replacement from the list of all species in the phylogeny pool. All species in the phylogeny will have equal probability of being included in the null communities. By changing the phylogeny, different species pools can be simulated. For example, the phylogeny could include the species present in some larger region.

3 - Independent swap: The independent swap algorithm (Gotelli and Entsminger, 2003); also known as ‘SIM9’ (Gotelli, 2000) creates swapped versions of the sample/species matrix.

Examples

sfile <- system.file("examples/sample_comstruct", package = "phylocomr")
pfile <- system.file("examples/phylo_comstruct", package = "phylocomr")

# from data.frame
sampledf <- read.table(sfile, header = FALSE,
  stringsAsFactors = FALSE)
phylo_str <- readLines(pfile)
ph_comdist(sample = sampledf, phylo = phylo_str)

#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 4.250000 6.749999 8.083335 8.708335 8.062500 8.046875
#> 2 clump2a 6.750001 4.944444 8.722219 8.416661 8.062505 7.822916
#> 3 clump2b 8.083340 8.722218 5.833330 7.361108 8.062505 7.947916
#> 4  clump4 8.708339 8.416662 7.361108 6.944441 7.875004 8.239582
#> 5    even 8.062500 8.062500 8.062496 7.874995 7.750000 8.000000
#> 6  random 8.046875 7.822917 7.947919 8.239586 8.000000 7.109375

ph_comdistnt(sample = sampledf, phylo = phylo_str)

#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4  even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl> <dbl>    <dbl>
#> 1  clump1 2.000000 4.166667 4.833334 6.000000 4.750 4.875000
#> 2 clump2a 4.166667 2.000000 6.000000 4.333333 4.500 4.625000
#> 3 clump2b 4.833333 6.000000 2.000000 3.000000 4.000 3.937500
#> 4  clump4 6.000000 4.333334 3.000000 2.000000 2.000 4.104167
#> 5    even 4.750000 4.500000 4.000000 2.000000 6.000 2.625000
#> 6  random 4.875000 4.625000 3.937500 4.104167 2.625 4.875000

# from files
sample_str <- paste0(readLines(sfile), collapse = "\n")
sfile2 <- tempfile()
cat(sample_str, file = sfile2, sep = '\n')
pfile2 <- tempfile()
cat(phylo_str, file = pfile2, sep = '\n')
ph_comdist(sample = sfile2, phylo = pfile2)

#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 4.250000 6.749999 8.083335 8.708335 8.062500 8.046875
#> 2 clump2a 6.750001 4.944444 8.722219 8.416661 8.062505 7.822916
#> 3 clump2b 8.083340 8.722218 5.833330 7.361108 8.062505 7.947916
#> 4  clump4 8.708339 8.416662 7.361108 6.944441 7.875004 8.239582
#> 5    even 8.062500 8.062500 8.062496 7.874995 7.750000 8.000000
#> 6  random 8.046875 7.822917 7.947919 8.239586 8.000000 7.109375

ph_comdistnt(sample = sfile2, phylo = pfile2)

#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4  even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl> <dbl>    <dbl>
#> 1  clump1 2.000000 4.166667 4.833334 6.000000 4.750 4.875000
#> 2 clump2a 4.166667 2.000000 6.000000 4.333333 4.500 4.625000
#> 3 clump2b 4.833333 6.000000 2.000000 3.000000 4.000 3.937500
#> 4  clump4 6.000000 4.333334 3.000000 2.000000 2.000 4.104167
#> 5    even 4.750000 4.500000 4.000000 2.000000 6.000 2.625000
#> 6  random 4.875000 4.625000 3.937500 4.104167 2.625 4.875000