comdist

Outputs the phylogenetic distance between samples, based on phylogenetic distances of taxa in one sample to the taxa in the other

ph_comdist(sample, phylo, rand_test = FALSE, null_model = 0,
  randomizations = 999, abundance = TRUE)

ph_comdistnt(sample, phylo, rand_test = FALSE, null_model = 0,
  randomizations = 999, abundance = TRUE)

Arguments

sample

(data.frame/character) sample data.frame or path to a sample file

phylo

(character/phylo) One of: phylogeny as a newick string (will be written to a temp file) - OR path to file with a newick string - OR a an ape `phylo` object. required.

rand_test

(logical) do you want to use null models? Default: FALSE

null_model

(integer) which null model to use. See Details.

randomizations

(numeric) number of randomizations. Default: 999

abundance

(logical) If TRUE (default) computed accounting for abundance. Otherwise, uses presence-absence.

Value

data.frame or a list of data.frame's if use null models

Null models

0 - Phylogeny shuffle: This null model shuffles species labels across the entire phylogeny. This randomizes phylogenetic relationships among species.

1 - Species in each sample become random draws from sample pool: This null model maintains the species richness of each sample, but the identities of the species occurring in each sample are randomized. For each sample, species are drawn without replacement from the list of all species actually occurring in at least one sample. Thus, species in the phylogeny that are not actually observed to occur in a sample will not be included in the null communities

2 - Species in each sample become random draws from phylogeny pool: This null model maintains the species richness of each sample, but the identities of the species occurring in each sample are randomized. For each sample, species are drawn without replacement from the list of all species in the phylogeny pool. All species in the phylogeny will have equal probability of being included in the null communities. By changing the phylogeny, different species pools can be simulated. For example, the phylogeny could include the species present in some larger region.

3 - Independent swap: The independent swap algorithm (Gotelli and Entsminger, 2003); also known as ‘SIM9’ (Gotelli, 2000) creates swapped versions of the sample/species matrix.

Examples

sfile <- system.file("examples/sample_comstruct", package = "phylocomr")
pfile <- system.file("examples/phylo_comstruct", package = "phylocomr")

# from data.frame
sampledf <- read.table(sfile, header = FALSE,
  stringsAsFactors = FALSE)
phylo_str <- readLines(pfile)
ph_comdist(sample = sampledf, phylo = phylo_str)

#> # A tibble: 6 × 7
#>      name   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 4.250000 6.749999 8.083335 8.708335 8.062500 8.046875
#> 2 clump2a 6.750001 4.944444 8.722219 8.416661 8.062505 7.822916
#> 3 clump2b 8.083340 8.722218 5.833330 7.361108 8.062505 7.947916
#> 4  clump4 8.708339 8.416662 7.361108 6.944441 7.875004 8.239582
#> 5    even 8.062500 8.062500 8.062496 7.874995 7.750000 8.000000
#> 6  random 8.046875 7.822917 7.947919 8.239586 8.000000 7.109375

ph_comdistnt(sample = sampledf, phylo = phylo_str)

#> # A tibble: 6 × 7
#>      name   clump1  clump2a  clump2b   clump4  even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl> <dbl>    <dbl>
#> 1  clump1 2.000000 4.166667 4.833334 6.000000 4.750 4.875000
#> 2 clump2a 4.166667 2.000000 6.000000 4.333333 4.500 4.625000
#> 3 clump2b 4.833333 6.000000 2.000000 3.000000 4.000 3.937500
#> 4  clump4 6.000000 4.333334 3.000000 2.000000 2.000 4.104167
#> 5    even 4.750000 4.500000 4.000000 2.000000 6.000 2.625000
#> 6  random 4.875000 4.625000 3.937500 4.104167 2.625 4.875000

ph_comdist(sample = sampledf, phylo = phylo_str, rand_test = TRUE)

#> $obs
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 4.250000 6.749999 8.083335 8.708335 8.062500 8.046875
#> 2 clump2a 6.750001 4.944444 8.722219 8.416661 8.062505 7.822916
#> 3 clump2b 8.083340 8.722218 5.833330 7.361108 8.062505 7.947916
#> 4  clump4 8.708339 8.416662 7.361108 6.944441 7.875004 8.239582
#> 5    even 8.062500 8.062500 8.062496 7.874995 7.750000 8.000000
#> 6  random 8.046875 7.822917 7.947919 8.239586 8.000000 7.109375
#> 
#> $null_mean
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 7.269207 7.966617 7.960646 8.153592 8.061437 8.111486
#> 2 clump2a 7.966617 7.150069 8.090708 7.968038 8.057477 8.187826
#> 3 clump2b 7.960646 8.090708 7.178596 7.758713 8.061184 8.009711
#> 4  clump4 8.153592 7.968038 7.758713 7.181819 7.815234 8.019092
#> 5    even 8.061437 8.057477 8.061184 7.815234 7.297235 8.013169
#> 6  random 8.111486 8.187826 8.009711 8.019092 8.013169 7.032251
#> 
#> $null_sd
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 0.295959 0.230404 0.220407 0.212135 0.198410 0.235767
#> 2 clump2a 0.230404 0.322466 0.236113 0.230751 0.209572 0.264318
#> 3 clump2b 0.220407 0.236113 0.300510 0.236081 0.214902 0.242678
#> 4  clump4 0.212135 0.230751 0.236081 0.323953 0.215324 0.234125
#> 5    even 0.198410 0.209572 0.214902 0.215324 0.275008 0.218760
#> 6  random 0.235767 0.264318 0.242678 0.234125 0.218760 0.334773
#> 
#> $NRI_or_NTI
#>      name    clump1   clump2a   clump2b    clump4      even    random
#> 1  clump1 10.201423  5.280374 -0.556647 -2.615048 -0.005359  0.274048
#> 2 clump2a  5.280361  6.839875 -2.674611 -1.944186 -0.023991  1.380574
#> 3 clump2b -0.556668 -2.674607  4.476611  1.684186 -0.006146  0.254641
#> 4  clump4 -2.615066 -1.944190  1.684184  0.732756 -0.277584 -0.941764
#> 5    even -0.005359 -0.023968 -0.006106 -0.277542 -1.646371  0.060200
#> 6  random  0.274048  1.380573  0.254627 -0.941781  0.060200 -0.230377
#> 

ph_comdistnt(sample = sampledf, phylo = phylo_str, rand_test = TRUE)

#> $obs
#>      name   clump1  clump2a  clump2b   clump4  even   random
#> 1  clump1 2.000000 4.166667 4.833334 6.000000 4.750 4.875000
#> 2 clump2a 4.166667 2.000000 6.000000 4.333333 4.500 4.625000
#> 3 clump2b 4.833333 6.000000 2.000000 3.000000 4.000 3.937500
#> 4  clump4 6.000000 4.333334 3.000000 2.000000 2.000 4.104167
#> 5    even 4.750000 4.500000 4.000000 2.000000 6.000 2.625000
#> 6  random 4.875000 4.625000 3.937500 4.104167 2.625 4.875000
#> 
#> $null_mean
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 4.714465 2.694694 2.655356 3.545245 3.327828 3.509093
#> 2 clump2a 2.694694 4.696031 2.987152 2.589090 3.330531 3.664309
#> 3 clump2b 2.655356 2.987152 4.737572 2.234985 3.333931 3.321749
#> 4  clump4 3.545245 2.589090 2.234985 4.725061 2.232033 3.341915
#> 5    even 3.327828 3.330531 3.333931 2.232033 4.728729 3.163748
#> 6  random 3.509093 3.664309 3.321749 3.341915 3.163748 4.742868
#> 
#> $null_sd
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 0.637319 0.484905 0.472669 0.578813 0.528240 0.570263
#> 2 clump2a 0.484905 0.683876 0.521981 0.444919 0.547392 0.621409
#> 3 clump2b 0.472669 0.521981 0.688280 0.398638 0.554559 0.561783
#> 4  clump4 0.578813 0.444919 0.398638 0.698461 0.383496 0.553392
#> 5    even 0.528240 0.547392 0.554559 0.383496 0.666331 0.546922
#> 6  random 0.570263 0.621409 0.561783 0.553392 0.546922 0.759870
#> 
#> $NRI_or_NTI
#>      name    clump1   clump2a   clump2b    clump4      even    random
#> 1  clump1  4.259195 -3.035591 -4.607831 -4.241020 -2.692282 -2.395224
#> 2 clump2a -3.035591  3.942282 -5.771955 -3.920363 -2.136436 -1.545989
#> 3 clump2b -4.607830 -5.771953  3.977408 -1.919073 -1.201079 -1.096066
#> 4  clump4 -4.241020 -3.920364 -1.919072  3.901523  0.605046 -1.377419
#> 5    even -2.692282 -2.136437 -1.201079  0.605046 -1.907867  0.985054
#> 6  random -2.395224 -1.545989 -1.096066 -1.377419  0.985054 -0.173888
#> 

# from files
sample_str <- paste0(readLines(sfile), collapse = "\n")
sfile2 <- tempfile()
cat(sample_str, file = sfile2, sep = '\n')
pfile2 <- tempfile()
cat(phylo_str, file = pfile2, sep = '\n')
ph_comdist(sample = sfile2, phylo = pfile2)

#> # A tibble: 6 × 7
#>      name   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 4.250000 6.749999 8.083335 8.708335 8.062500 8.046875
#> 2 clump2a 6.750001 4.944444 8.722219 8.416661 8.062505 7.822916
#> 3 clump2b 8.083340 8.722218 5.833330 7.361108 8.062505 7.947916
#> 4  clump4 8.708339 8.416662 7.361108 6.944441 7.875004 8.239582
#> 5    even 8.062500 8.062500 8.062496 7.874995 7.750000 8.000000
#> 6  random 8.046875 7.822917 7.947919 8.239586 8.000000 7.109375

ph_comdistnt(sample = sfile2, phylo = pfile2)

#> # A tibble: 6 × 7
#>      name   clump1  clump2a  clump2b   clump4  even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl> <dbl>    <dbl>
#> 1  clump1 2.000000 4.166667 4.833334 6.000000 4.750 4.875000
#> 2 clump2a 4.166667 2.000000 6.000000 4.333333 4.500 4.625000
#> 3 clump2b 4.833333 6.000000 2.000000 3.000000 4.000 3.937500
#> 4  clump4 6.000000 4.333334 3.000000 2.000000 2.000 4.104167
#> 5    even 4.750000 4.500000 4.000000 2.000000 6.000 2.625000
#> 6  random 4.875000 4.625000 3.937500 4.104167 2.625 4.875000

ph_comdist(sample = sfile2, phylo = pfile2, rand_test = TRUE)

#> $obs
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 4.250000 6.749999 8.083335 8.708335 8.062500 8.046875
#> 2 clump2a 6.750001 4.944444 8.722219 8.416661 8.062505 7.822916
#> 3 clump2b 8.083340 8.722218 5.833330 7.361108 8.062505 7.947916
#> 4  clump4 8.708339 8.416662 7.361108 6.944441 7.875004 8.239582
#> 5    even 8.062500 8.062500 8.062496 7.874995 7.750000 8.000000
#> 6  random 8.046875 7.822917 7.947919 8.239586 8.000000 7.109375
#> 
#> $null_mean
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 7.269207 7.966617 7.960646 8.153592 8.061437 8.111486
#> 2 clump2a 7.966617 7.150069 8.090708 7.968038 8.057477 8.187826
#> 3 clump2b 7.960646 8.090708 7.178596 7.758713 8.061184 8.009711
#> 4  clump4 8.153592 7.968038 7.758713 7.181819 7.815234 8.019092
#> 5    even 8.061437 8.057477 8.061184 7.815234 7.297235 8.013169
#> 6  random 8.111486 8.187826 8.009711 8.019092 8.013169 7.032251
#> 
#> $null_sd
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 0.295959 0.230404 0.220407 0.212135 0.198410 0.235767
#> 2 clump2a 0.230404 0.322466 0.236113 0.230751 0.209572 0.264318
#> 3 clump2b 0.220407 0.236113 0.300510 0.236081 0.214902 0.242678
#> 4  clump4 0.212135 0.230751 0.236081 0.323953 0.215324 0.234125
#> 5    even 0.198410 0.209572 0.214902 0.215324 0.275008 0.218760
#> 6  random 0.235767 0.264318 0.242678 0.234125 0.218760 0.334773
#> 
#> $NRI_or_NTI
#>      name    clump1   clump2a   clump2b    clump4      even    random
#> 1  clump1 10.201423  5.280374 -0.556647 -2.615048 -0.005359  0.274048
#> 2 clump2a  5.280361  6.839875 -2.674611 -1.944186 -0.023991  1.380574
#> 3 clump2b -0.556668 -2.674607  4.476611  1.684186 -0.006146  0.254641
#> 4  clump4 -2.615066 -1.944190  1.684184  0.732756 -0.277584 -0.941764
#> 5    even -0.005359 -0.023968 -0.006106 -0.277542 -1.646371  0.060200
#> 6  random  0.274048  1.380573  0.254627 -0.941781  0.060200 -0.230377
#> 

ph_comdistnt(sample = sfile2, phylo = pfile2, rand_test = TRUE)

#> $obs
#>      name   clump1  clump2a  clump2b   clump4  even   random
#> 1  clump1 2.000000 4.166667 4.833334 6.000000 4.750 4.875000
#> 2 clump2a 4.166667 2.000000 6.000000 4.333333 4.500 4.625000
#> 3 clump2b 4.833333 6.000000 2.000000 3.000000 4.000 3.937500
#> 4  clump4 6.000000 4.333334 3.000000 2.000000 2.000 4.104167
#> 5    even 4.750000 4.500000 4.000000 2.000000 6.000 2.625000
#> 6  random 4.875000 4.625000 3.937500 4.104167 2.625 4.875000
#> 
#> $null_mean
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 4.714465 2.694694 2.655356 3.545245 3.327828 3.509093
#> 2 clump2a 2.694694 4.696031 2.987152 2.589090 3.330531 3.664309
#> 3 clump2b 2.655356 2.987152 4.737572 2.234985 3.333931 3.321749
#> 4  clump4 3.545245 2.589090 2.234985 4.725061 2.232033 3.341915
#> 5    even 3.327828 3.330531 3.333931 2.232033 4.728729 3.163748
#> 6  random 3.509093 3.664309 3.321749 3.341915 3.163748 4.742868
#> 
#> $null_sd
#>      name   clump1  clump2a  clump2b   clump4     even   random
#> 1  clump1 0.637319 0.484905 0.472669 0.578813 0.528240 0.570263
#> 2 clump2a 0.484905 0.683876 0.521981 0.444919 0.547392 0.621409
#> 3 clump2b 0.472669 0.521981 0.688280 0.398638 0.554559 0.561783
#> 4  clump4 0.578813 0.444919 0.398638 0.698461 0.383496 0.553392
#> 5    even 0.528240 0.547392 0.554559 0.383496 0.666331 0.546922
#> 6  random 0.570263 0.621409 0.561783 0.553392 0.546922 0.759870
#> 
#> $NRI_or_NTI
#>      name    clump1   clump2a   clump2b    clump4      even    random
#> 1  clump1  4.259195 -3.035591 -4.607831 -4.241020 -2.692282 -2.395224
#> 2 clump2a -3.035591  3.942282 -5.771955 -3.920363 -2.136436 -1.545989
#> 3 clump2b -4.607830 -5.771953  3.977408 -1.919073 -1.201079 -1.096066
#> 4  clump4 -4.241020 -3.920364 -1.919072  3.901523  0.605046 -1.377419
#> 5    even -2.692282 -2.136437 -1.201079  0.605046 -1.907867  0.985054
#> 6  random -2.395224 -1.545989 -1.096066 -1.377419  0.985054 -0.173888
#>