comstruct

Calculates mean phylogenetic distance (MPD) and mean nearest phylogenetic taxon distance (MNTD; aka MNND) for each sample, and compares them to MPD/MNTD values for randomly generated samples (null communities) or phylogenies.

ph_comstruct(sample, phylo, null_model = 0, randomizations = 999,
  swaps = 1000, abundance = TRUE)

Arguments

sample

(data.frame/character) sample data.frame or path to a sample file

phylo

(character/phylo) One of: phylogeny as a newick string (will be written to a temp file) - OR path to file with a newick string - OR a an ape `phylo` object. required.

null_model

(integer) which null model to use. See Details.

randomizations

(numeric) number of randomizations. Default: 999

swaps

(numeric) number of swaps. Default: 1000

abundance

(logical) If TRUE (default) computed accounting for abundance. Otherwise, uses presence-absence.

Value

data.frame

Null models

0 - Phylogeny shuffle: This null model shuffles species labels across the entire phylogeny. This randomizes phylogenetic relationships among species.

1 - Species in each sample become random draws from sample pool: This null model maintains the species richness of each sample, but the identities of the species occurring in each sample are randomized. For each sample, species are drawn without replacement from the list of all species actually occurring in at least one sample. Thus, species in the phylogeny that are not actually observed to occur in a sample will not be included in the null communities

2 - Species in each sample become random draws from phylogeny pool: This null model maintains the species richness of each sample, but the identities of the species occurring in each sample are randomized. For each sample, species are drawn without replacement from the list of all species in the phylogeny pool. All species in the phylogeny will have equal probability of being included in the null communities. By changing the phylogeny, different species pools can be simulated. For example, the phylogeny could include the species present in some larger region.

3 - Independent swap: The independent swap algorithm (Gotelli and Entsminger, 2003); also known as ‘SIM9’ (Gotelli, 2000) creates swapped versions of the sample/species matrix.

Examples

sfile <- system.file("examples/sample_comstruct", package = "phylocomr")
pfile <- system.file("examples/phylo_comstruct", package = "phylocomr")

# from data.frame
sampledf <- read.table(sfile, header = FALSE,
  stringsAsFactors = FALSE)
phylo_str <- readLines(pfile)
(res <- ph_comstruct(sample = sampledf, phylo = phylo_str))

#> # A tibble: 6 × 15
#>      plot ntaxa    mpd mpd.rnd mpd.sd     nri mpd.ranklow mpd.rankhi  mntd
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.2692 0.2960 10.2014         999          0 2.000
#> 2 clump2a     8 4.9444  7.1501 0.3225  6.8399         999          0 2.000
#> 3 clump2b     8 5.8333  7.1786 0.3005  4.4766         996          3 2.000
#> 4  clump4     8 6.9444  7.1818 0.3240  0.7327         824        189 2.000
#> 5    even     8 7.7500  7.2972 0.2750 -1.6464           8        999 6.000
#> 6  random     8 7.1094  7.0323 0.3348 -0.2304         500        509 4.875
#> # ... with 6 more variables: mntd.rnd <dbl>, mntd.sd <dbl>, nti <dbl>,
#> #   mntd.ranklo <int>, mntd.rankhi <int>, runs <int>

# from files
sample_str <- paste0(readLines(sfile), collapse = "\n")
sfile2 <- tempfile()
cat(sample_str, file = sfile2, sep = '\n')
pfile2 <- tempfile()
cat(phylo_str, file = pfile2, sep = '\n')
(res <- ph_comstruct(sample = sfile2, phylo = pfile2))

#> # A tibble: 6 × 15
#>      plot ntaxa    mpd mpd.rnd mpd.sd     nri mpd.ranklow mpd.rankhi  mntd
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.2692 0.2960 10.2014         999          0 2.000
#> 2 clump2a     8 4.9444  7.1501 0.3225  6.8399         999          0 2.000
#> 3 clump2b     8 5.8333  7.1786 0.3005  4.4766         996          3 2.000
#> 4  clump4     8 6.9444  7.1818 0.3240  0.7327         824        189 2.000
#> 5    even     8 7.7500  7.2972 0.2750 -1.6464           8        999 6.000
#> 6  random     8 7.1094  7.0323 0.3348 -0.2304         500        509 4.875
#> # ... with 6 more variables: mntd.rnd <dbl>, mntd.sd <dbl>, nti <dbl>,
#> #   mntd.ranklo <int>, mntd.rankhi <int>, runs <int>

# different null models
ph_comstruct(sample = sfile2, phylo = pfile2, null_model = 0)

#> # A tibble: 6 × 15
#>      plot ntaxa    mpd mpd.rnd mpd.sd     nri mpd.ranklow mpd.rankhi  mntd
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.2692 0.2960 10.2014         999          0 2.000
#> 2 clump2a     8 4.9444  7.1501 0.3225  6.8399         999          0 2.000
#> 3 clump2b     8 5.8333  7.1786 0.3005  4.4766         996          3 2.000
#> 4  clump4     8 6.9444  7.1818 0.3240  0.7327         824        189 2.000
#> 5    even     8 7.7500  7.2972 0.2750 -1.6464           8        999 6.000
#> 6  random     8 7.1094  7.0323 0.3348 -0.2304         500        509 4.875
#> # ... with 6 more variables: mntd.rnd <dbl>, mntd.sd <dbl>, nti <dbl>,
#> #   mntd.ranklo <int>, mntd.rankhi <int>, runs <int>

ph_comstruct(sample = sfile2, phylo = pfile2, null_model = 1)

#> # A tibble: 6 × 15
#>      plot ntaxa    mpd mpd.rnd mpd.sd     nri mpd.ranklow mpd.rankhi  mntd
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.0685 0.4209  6.6960         999          0 2.000
#> 2 clump2a     8 4.9444  7.0701 0.3745  5.6757         999          0 2.000
#> 3 clump2b     8 5.8333  7.0923 0.3693  3.4092         991          9 2.000
#> 4  clump4     8 6.9444  7.0778 0.3578  0.3727         737        290 2.000
#> 5    even     8 7.7500  7.2082 0.3148 -1.7212           7        999 6.000
#> 6  random     8 7.1094  6.9559 0.3844 -0.3991         433        583 4.875
#> # ... with 6 more variables: mntd.rnd <dbl>, mntd.sd <dbl>, nti <dbl>,
#> #   mntd.ranklo <int>, mntd.rankhi <int>, runs <int>

ph_comstruct(sample = sfile2, phylo = pfile2, null_model = 2)

#> # A tibble: 6 × 15
#>      plot ntaxa    mpd mpd.rnd mpd.sd     nri mpd.ranklow mpd.rankhi  mntd
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.2818 0.2910 10.4171         999          0 2.000
#> 2 clump2a     8 4.9444  7.1708 0.3135  7.1017         999          0 2.000
#> 3 clump2b     8 5.8333  7.1624 0.3144  4.2280         997          2 2.000
#> 4  clump4     8 6.9444  7.1667 0.3321  0.6691         813        210 2.000
#> 5    even     8 7.7500  7.2842 0.2813 -1.6557           5        999 6.000
#> 6  random     8 7.1094  7.0160 0.3618 -0.2582         482        535 4.875
#> # ... with 6 more variables: mntd.rnd <dbl>, mntd.sd <dbl>, nti <dbl>,
#> #   mntd.ranklo <int>, mntd.rankhi <int>, runs <int>

# ph_comstruct(sample = sfile2, phylo = pfile2, null_model = 3)