comstruct

comstruct

ph_comstruct(sample, phylo, null_model = 0, randomizations = 999,
  swaps = 1000, abundance = TRUE)

Arguments

sample

(data.frame/character) sample data.frame or path to a sample file

phylo

(character/phylo) phylogeny as a phylo object or a newick string (will be written to a temp file if provided) - or a path to a file with a newick string

null_model

(integer) which null model to use. See Details.

randomizations

(numeric) number of randomizations. Default: 999

swaps

(numeric) number of swaps. Default: 1000

abundance

(logical) If TRUE (default) computed accounting for abundance. Otherwise, uses presence-absence.

Value

data.frame

Null models

0 - Phylogeny shuffle: This null model shuffles species labels across the entire phylogeny. This randomizes phylogenetic relationships among species.

1 - Species in each sample become random draws from sample pool: This null model maintains the species richness of each sample, but the identities of the species occurring in each sample are randomized. For each sample, species are drawn without replacement from the list of all species actually occurring in at least one sample. Thus, species in the phylogeny that are not actually observed to occur in a sample will not be included in the null communities

2 - Species in each sample become random draws from phylogeny pool: This null model maintains the species richness of each sample, but the identities of the species occurring in each sample are randomized. For each sample, species are drawn without replacement from the list of all species in the phylogeny pool. All species in the phylogeny will have equal probability of being included in the null communities. By changing the phylogeny, different species pools can be simulated. For example, the phylogeny could include the species present in some larger region.

3 - Independent swap: The independent swap algorithm (Gotelli and Entsminger, 2003); also known as ‘SIM9’ (Gotelli, 2000) creates swapped versions of the sample/species matrix.

Examples

sfile <- system.file("examples/sample_comstruct", package = "phylocomr")
pfile <- system.file("examples/phylo_comstruct", package = "phylocomr")

# from data.frame
sampledf <- read.table(sfile, header = FALSE,
  stringsAsFactors = FALSE)
phylo_str <- readLines(pfile)
(res <- ph_comstruct(sample = sampledf, phylo = phylo_str))

#> # A tibble: 6 × 15
#>      plot ntaxa    MPD MPD.rnd MPD.sd     NRI MPD.rankLow MPD.rankHi  MNTD
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.2675 0.3060  9.8606         999          0 2.000
#> 2 clump2a     8 4.9444  7.1575 0.3205  6.9053         999          0 2.000
#> 3 clump2b     8 5.8333  7.1565 0.3172  4.1712         993          6 2.000
#> 4  clump4     8 6.9444  7.1668 0.3101  0.7172         815        206 2.000
#> 5    even     8 7.7500  7.2840 0.2837 -1.6428           6        999 6.000
#> 6  random     8 7.1094  7.0061 0.3462 -0.2985         460        554 4.875
#> # ... with 6 more variables: MNTD.rnd <dbl>, MNTD.sd <dbl>, NTI <dbl>,
#> #   MNTD.rankLo <int>, MNTD.rankHi <int>, runs <int>

# from files
sample_str <- paste0(readLines(sfile), collapse = "\n")
sfile2 <- tempfile()
cat(sample_str, file = sfile2, sep = '\n')
pfile2 <- tempfile()
cat(phylo_str, file = pfile2, sep = '\n')
(res <- ph_comstruct(sample = sfile2, phylo = pfile2))

#> # A tibble: 6 × 15
#>      plot ntaxa    MPD MPD.rnd MPD.sd     NRI MPD.rankLow MPD.rankHi  MNTD
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.2675 0.3060  9.8606         999          0 2.000
#> 2 clump2a     8 4.9444  7.1575 0.3205  6.9053         999          0 2.000
#> 3 clump2b     8 5.8333  7.1565 0.3172  4.1712         993          6 2.000
#> 4  clump4     8 6.9444  7.1668 0.3101  0.7172         815        206 2.000
#> 5    even     8 7.7500  7.2840 0.2837 -1.6428           6        999 6.000
#> 6  random     8 7.1094  7.0061 0.3462 -0.2985         460        554 4.875
#> # ... with 6 more variables: MNTD.rnd <dbl>, MNTD.sd <dbl>, NTI <dbl>,
#> #   MNTD.rankLo <int>, MNTD.rankHi <int>, runs <int>

# different null models
ph_comstruct(sample = sfile2, phylo = pfile2, null_model = 0)

#> # A tibble: 6 × 15
#>      plot ntaxa    MPD MPD.rnd MPD.sd     NRI MPD.rankLow MPD.rankHi  MNTD
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.2675 0.3060  9.8606         999          0 2.000
#> 2 clump2a     8 4.9444  7.1575 0.3205  6.9053         999          0 2.000
#> 3 clump2b     8 5.8333  7.1565 0.3172  4.1712         993          6 2.000
#> 4  clump4     8 6.9444  7.1668 0.3101  0.7172         815        206 2.000
#> 5    even     8 7.7500  7.2840 0.2837 -1.6428           6        999 6.000
#> 6  random     8 7.1094  7.0061 0.3462 -0.2985         460        554 4.875
#> # ... with 6 more variables: MNTD.rnd <dbl>, MNTD.sd <dbl>, NTI <dbl>,
#> #   MNTD.rankLo <int>, MNTD.rankHi <int>, runs <int>

ph_comstruct(sample = sfile2, phylo = pfile2, null_model = 1)

#> # A tibble: 6 × 15
#>      plot ntaxa    MPD MPD.rnd MPD.sd     NRI MPD.rankLow MPD.rankHi  MNTD
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.0913 0.4003  7.0975         999          0 2.000
#> 2 clump2a     8 4.9444  7.0962 0.3556  6.0511         999          0 2.000
#> 3 clump2b     8 5.8333  7.0900 0.3491  3.5992         991          9 2.000
#> 4  clump4     8 6.9444  7.0930 0.3365  0.4416         748        274 2.000
#> 5    even     8 7.7500  7.2055 0.3067 -1.7755           3        999 6.000
#> 6  random     8 7.1094  6.9310 0.3936 -0.4530         398        614 4.875
#> # ... with 6 more variables: MNTD.rnd <dbl>, MNTD.sd <dbl>, NTI <dbl>,
#> #   MNTD.rankLo <int>, MNTD.rankHi <int>, runs <int>

ph_comstruct(sample = sfile2, phylo = pfile2, null_model = 2)

#> # A tibble: 6 × 15
#>      plot ntaxa    MPD MPD.rnd MPD.sd     NRI MPD.rankLow MPD.rankHi  MNTD
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  7.2934 0.2716 11.2059         999          0 2.000
#> 2 clump2a     8 4.9444  7.1540 0.3118  7.0856         999          0 2.000
#> 3 clump2b     8 5.8333  7.1740 0.3140  4.2693         996          3 2.000
#> 4  clump4     8 6.9444  7.1575 0.3349  0.6362         813        208 2.000
#> 5    even     8 7.7500  7.2762 0.3002 -1.5783           8        999 6.000
#> 6  random     8 7.1094  7.0248 0.3351 -0.2525         478        544 4.875
#> # ... with 6 more variables: MNTD.rnd <dbl>, MNTD.sd <dbl>, NTI <dbl>,
#> #   MNTD.rankLo <int>, MNTD.rankHi <int>, runs <int>

ph_comstruct(sample = sfile2, phylo = pfile2, null_model = 3)

#> # A tibble: 6 × 15
#>      plot ntaxa    MPD MPD.rnd MPD.sd     NRI MPD.rankLow MPD.rankHi  MNTD
#>     <chr> <int>  <dbl>   <dbl>  <dbl>   <dbl>       <int>      <int> <dbl>
#> 1  clump1     8 4.2500  6.9381 0.3722  7.2227         999          0 2.000
#> 2 clump2a     8 4.9444  6.9486 0.3615  5.5437         999          0 2.000
#> 3 clump2b     8 5.8333  6.9477 0.3786  2.9434         983         16 2.000
#> 4  clump4     8 6.9444  6.9309 0.3800 -0.0357         557        447 2.000
#> 5    even     8 7.7500  6.9488 0.3721 -2.1532           0        999 6.000
#> 6  random     8 7.1094  6.9298 0.3952 -0.4544         370        629 4.875
#> # ... with 6 more variables: MNTD.rnd <dbl>, MNTD.sd <dbl>, NTI <dbl>,
#> #   MNTD.rankLo <int>, MNTD.rankHi <int>, runs <int>