ecovolve

Ecovolve generates a phylogeny via a random birth and death process, generates a traits file with five randomly evolving, in-dependent traits, and a sample file with a single sample unit (‘alive’) containing all extant members of the phylogeny.

ph_ecovolve(speciation = 0.05, extinction = 0.01, time_units = 100,
  out_mode = 3, prob_env = "3211000000", extant_lineages = FALSE,
  only_extant = FALSE, taper_change = NULL, competition = FALSE)

Arguments

speciation

(numeric) Probability of speciation per unit time. Default: 0.05

extinction

(numeric) Probability of extinction per unit time. Default: 0.01

time_units

(integer) Time units to simulate over. Default: 100

out_mode

(integer) Output mode (2 = LTT; 3 = newick). Default: 3

prob_env

(character) Probability envelope for character change. must be a string of 10 integers. Default: 3211000000

extant_lineages

(logical) Stop simulation after this number of extant lineages. Default: FALSE

only_extant

(logical) Output phylogeny pruned only for extant taxa. Default: FALSE

taper_change

(numeric/integer) Taper character change by e^(-time/F). This produces more conservatism in traits (see Kraft et al., 2007). Default: NULL, not passed

competition

(logical) Simulate competition, with trait proximity increasing extinction. Default: FALSE

Value

a list with three elements:

phylogeny - a phylogeny as a newick string. In the case of out_mode = 2 gives a Lineage Through Time data.frame instead of a newick phylogeny

sample - a data.frame with three columns, "sample" (all "alive"), "abundance" (all 1's), "name" (the species code). In the case of out_mode = 2 gives an empty data.frame

traits - a data.frame with first column with spcies code ("name"), then 5 randomly evolved and independent traits. In the case of out_mode = 2 gives an empty data.frame

Clean up

Two files, "ecovolve.sample" and "ecovolve.traits" are written to the current working directory when this function runs - we read these files in, then delete the files via unlink

Failure behavior

Function occasionally fails with error "call to 'ecovolve' failed with status 8. only 1 taxon; > 1 required" - this just means that only 1 taxon was created in the random process, so the function can't proceed

Examples

# ph_ecovolve(speciation = 0.05)
# ph_ecovolve(speciation = 0.1)
ph_ecovolve(extinction = 0.005)

#> $phylogeny
#> [1] "(((dead28:17.000000,((sp220:20.000000,sp221:20.000000)node88:1.000000,(sp239:16.000000,sp240:16.000000)node108:5.000000)node79:36.000000)node10:18.000000,((((dead40:12.000000,(sp250:13.000000,(sp277:9.000000,sp278:9.000000)node141:4.000000)node121:33.000000)node22:10.000000,(sp203:30.000000,(sp216:20.000000,(sp229:18.000000,sp230:18.000000)node98:2.000000)node81:10.000000)node50:26.000000)node12:2.000000,((sp232:18.000000,sp233:18.000000)node100:1.000000,(sp280:8.000000,(sp295:7.000000,((sp310:5.000000,(sp319:4.000000,sp320:4.000000)node175:1.000000)node169:1.000000,(sp347:1.000000,sp348:1.000000)node190:5.000000)node162:1.000000)node157:1.000000)node145:11.000000)node89:39.000000)node9:14.000000,(((dead48:13.000000,sp194:44.000000)node24:3.000000,(((dead85:3.000000,(dead106:2.000000,sp226:19.000000)node94:4.000000)node72:7.000000,((dead147:8.000000,sp241:16.000000)node109:5.000000,sp215:21.000000)node80:9.000000)node51:10.000000,(dead117:15.000000,(sp222:19.000000,sp223:19.000000)node90:10.000000)node52:11.000000)node29:7.000000)node21:4.000000,(sp210:24.000000,(sp245:15.000000,(sp301:5.000000,sp302:5.000000)node164:10.000000)node113:9.000000)node67:27.000000)node15:21.000000)node3:3.000000)node1:4.000000,(((((dead35:12.000000,(sp198:35.000000,(sp243:15.000000,sp244:15.000000)node112:20.000000)node39:14.000000)node19:5.000000,((sp209:25.000000,((sp293:7.000000,sp294:7.000000)node156:1.000000,(sp306:5.000000,sp307:5.000000)node167:3.000000)node144:17.000000)node66:3.000000,(((sp246:15.000000,(sp321:3.000000,sp322:3.000000)node176:12.000000)node114:5.000000,(sp339:1.000000,sp340:1.000000)node186:19.000000)node87:2.000000,(sp312:4.000000,(sp330:3.000000,sp331:3.000000)node181:1.000000)node171:18.000000)node75:6.000000)node54:26.000000)node14:1.000000,(((dead110:2.000000,((sp279:9.000000,(sp323:3.000000,sp324:3.000000)node177:6.000000)node143:4.000000,((sp291:7.000000,(sp337:2.000000,sp338:2.000000)node185:5.000000)node154:2.000000,(sp292:7.000000,(sp308:5.000000,sp309:5.000000)node168:2.000000)node155:2.000000)node142:4.000000)node122:5.000000)node96:15.000000,sp199:33.000000)node42:5.000000,sp196:38.000000)node30:17.000000)node13:3.000000,(dead158:51.000000,((sp201:32.000000,(sp236:17.000000,sp237:17.000000)node104:15.000000)node45:9.000000,(sp247:14.000000,(sp255:12.000000,(sp298:6.000000,(sp327:3.000000,sp328:3.000000)node179:3.000000)node161:6.000000)node127:2.000000)node116:27.000000)node25:16.000000)node11:1.000000)node8:15.000000,(((((((dead59:2.000000,sp204:29.000000)node53:8.000000,(((dead83:6.000000,(sp224:19.000000,sp225:19.000000)node93:7.000000)node63:2.000000,((dead115:2.000000,(sp249:14.000000,(sp258:11.000000,(sp304:5.000000,sp305:5.000000)node166:6.000000)node130:3.000000)node119:3.000000)node105:5.000000,(sp299:5.000000,sp300:5.000000)node163:17.000000)node76:6.000000)node56:6.000000,(((dead152:12.000000,((sp284:8.000000,sp285:8.000000)node149:3.000000,(sp335:2.000000,sp336:2.000000)node184:9.000000)node129:8.000000)node92:7.000000,sp207:26.000000)node62:6.000000,(sp296:6.000000,sp297:6.000000)node159:26.000000)node46:2.000000)node41:3.000000)node36:1.000000,((sp219:20.000000,(sp248:14.000000,(sp290:7.000000,(sp325:3.000000,sp326:3.000000)node178:4.000000)node153:7.000000)node118:6.000000)node84:5.000000,(sp334:2.000000,(sp351:1.000000,sp352:1.000000)node192:1.000000)node183:23.000000)node64:13.000000)node31:12.000000,(sp197:38.000000,((sp259:10.000000,sp260:10.000000)node131:16.000000,(sp268:9.000000,sp269:9.000000)node136:17.000000)node61:12.000000)node32:12.000000)node18:9.000000,(((dead78:2.000000,(sp288:7.000000,sp289:7.000000)node151:17.000000)node68:9.000000,(sp266:9.000000,sp267:9.000000)node135:24.000000)node44:4.000000,((sp208:25.000000,((sp227:19.000000,(sp263:10.000000,sp264:10.000000)node133:9.000000)node95:3.000000,(sp228:18.000000,(sp254:12.000000,(sp303:5.000000,(sp317:4.000000,sp318:4.000000)node174:1.000000)node165:7.000000)node126:6.000000)node97:4.000000)node77:3.000000)node65:6.000000,((sp211:24.000000,sp212:24.000000)node69:4.000000,(sp332:2.000000,sp333:2.000000)node182:26.000000)node57:3.000000)node49:6.000000)node34:22.000000)node7:11.000000,(sp193:48.000000,(sp202:31.000000,(sp286:7.000000,sp287:7.000000)node150:24.000000)node47:17.000000)node20:22.000000)node5:1.000000,(((((dead120:1.000000,sp242:15.000000)node111:21.000000,(sp311:4.000000,(sp329:3.000000,(sp349:1.000000,sp350:1.000000)node191:2.000000)node180:1.000000)node170:32.000000)node37:9.000000,(((sp213:24.000000,(sp270:9.000000,(sp315:4.000000,sp316:4.000000)node173:5.000000)node137:15.000000)node71:2.000000,(sp217:20.000000,sp218:20.000000)node82:6.000000)node60:15.000000,(((sp231:18.000000,(sp238:17.000000,((sp256:12.000000,sp257:12.000000)node128:1.000000,(sp283:8.000000,(sp345:1.000000,sp346:1.000000)node189:7.000000)node148:5.000000)node123:4.000000)node107:1.000000)node99:2.000000,(sp273:9.000000,sp274:9.000000)node139:11.000000)node86:2.000000,(sp261:10.000000,sp262:10.000000)node132:12.000000)node73:19.000000)node27:4.000000)node23:6.000000,(sp195:41.000000,((sp200:33.000000,(sp214:22.000000,(sp271:9.000000,sp272:9.000000)node138:13.000000)node74:11.000000)node43:5.000000,((sp275:9.000000,sp276:9.000000)node140:8.000000,(sp341:1.000000,sp342:1.000000)node187:16.000000)node102:21.000000)node33:3.000000)node26:10.000000)node16:8.000000,(((dead160:13.000000,((sp265:10.000000,(sp343:1.000000,sp344:1.000000)node188:9.000000)node134:8.000000,(sp281:8.000000,sp282:8.000000)node146:10.000000)node101:1.000000)node91:9.000000,sp205:28.000000)node55:23.000000,((sp206:27.000000,((sp251:12.000000,sp252:12.000000)node124:12.000000,(sp253:12.000000,(sp313:4.000000,sp314:4.000000)node172:8.000000)node125:12.000000)node70:3.000000)node58:8.000000,(sp234:17.000000,sp235:17.000000)node103:18.000000)node38:16.000000)node17:8.000000)node6:12.000000)node4:2.000000)node2:6.000000)node0:21.000000;\n"
#> 
#> $sample
#> # A tibble: 160 × 3
#>    sample abundance  name
#>     <chr>     <int> <chr>
#> 1   alive         1 sp193
#> 2   alive         1 sp194
#> 3   alive         1 sp195
#> 4   alive         1 sp196
#> 5   alive         1 sp197
#> 6   alive         1 sp198
#> 7   alive         1 sp199
#> 8   alive         1 sp200
#> 9   alive         1 sp201
#> 10  alive         1 sp202
#> # ... with 150 more rows
#> 
#> $traits
#> # A tibble: 177 × 6
#>       name ecov0 ecov1 ecov2 ecov3 ecov4
#>      <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   dead28     4    -2    18     5   -13
#> 2   dead35     1    -7     8    14     3
#> 3   dead40     4    18    20    -5    -2
#> 4   dead48   -13    20     8    -3     6
#> 5   dead59    13    -4     0     8   -11
#> 6   dead78     7    25    -7   -16   -21
#> 7   dead83     9    -3    -7     4    -9
#> 8   dead85   -23    26     9     7    -4
#> 9  dead106   -22    29    17     1    -6
#> 10 dead110     8    -7     6    -5    12
#> # ... with 167 more rows
#> 

ph_ecovolve(time_units = 50)

#> $phylogeny
#> [1] "((dead4:10.000000,((((dead21:6.000000,sp34:19.000000)node18:2.000000,dead23:10.000000)node16:7.000000,((sp31:23.000000,(sp36:16.000000,(sp37:12.000000,(sp45:4.000000,sp46:4.000000)node28:8.000000)node22:4.000000)node20:7.000000)node14:4.000000,(sp40:7.000000,(sp41:5.000000,sp42:5.000000)node26:2.000000)node25:20.000000)node10:1.000000)node8:9.000000,sp29:37.000000)node3:10.000000)node1:3.000000,(((dead9:4.000000,(sp30:24.000000,(sp35:18.000000,(sp38:10.000000,sp39:10.000000)node24:8.000000)node19:6.000000)node11:8.000000)node7:5.000000,(sp43:4.000000,sp44:4.000000)node27:33.000000)node5:5.000000,(dead12:9.000000,(dead15:1.000000,(sp32:19.000000,sp33:19.000000)node17:4.000000)node13:9.000000)node6:10.000000)node2:8.000000)node0:0.000000;\n"
#> 
#> $sample
#> # A tibble: 18 × 3
#>    sample abundance  name
#>     <chr>     <int> <chr>
#> 1   alive         1  sp29
#> 2   alive         1  sp30
#> 3   alive         1  sp31
#> 4   alive         1  sp32
#> 5   alive         1  sp33
#> 6   alive         1  sp34
#> 7   alive         1  sp35
#> 8   alive         1  sp36
#> 9   alive         1  sp37
#> 10  alive         1  sp38
#> 11  alive         1  sp39
#> 12  alive         1  sp40
#> 13  alive         1  sp41
#> 14  alive         1  sp42
#> 15  alive         1  sp43
#> 16  alive         1  sp44
#> 17  alive         1  sp45
#> 18  alive         1  sp46
#> 
#> $traits
#> # A tibble: 24 × 6
#>      name ecov0 ecov1 ecov2 ecov3 ecov4
#>     <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   dead4    -9     1    -2    10     2
#> 2   dead9     4     7    -6     2     4
#> 3  dead12    13     9    -6     8     7
#> 4  dead15     2    -1    -7     4     6
#> 5  dead21    -6     2    -2     8    14
#> 6  dead23   -11    -8    12    13     7
#> 7    sp29   -16    14    14    20    -3
#> 8    sp30     7     3    -7   -14   -14
#> 9    sp31    -4    17     2    -9    16
#> 10   sp32     2    -3    -2    11    15
#> # ... with 14 more rows
#> 

ph_ecovolve(out_mode = 2)

#> $phylogeny
#> # A tibble: 101 × 7
#>       v1    v2    v3    v4    v5    v6    v7
#>    <int> <int> <dbl> <dbl> <int> <dbl> <dbl>
#> 1      0     1     0   NaN     1     0   NaN
#> 2      1     1    -1   NaN     1    -1   NaN
#> 3      2     1     2   NaN     1     2   NaN
#> 4      3     1     1   NaN     1     1   NaN
#> 5      4     1     0   NaN     1     0   NaN
#> 6      5     1     1   NaN     1     1   NaN
#> 7      6     1     3   NaN     1     3   NaN
#> 8      7     1     5   NaN     1     5   NaN
#> 9      8     1     4   NaN     1     4   NaN
#> 10     9     1     3   NaN     1     3   NaN
#> # ... with 91 more rows
#> 
#> $sample
#> # A tibble: 0 × 0
#> 
#> $traits
#> # A tibble: 0 × 0
#> 

ph_ecovolve(extant_lineages = TRUE)

#> $phylogeny
#> [1] "((dead2:3.000000,(dead4:12.000000,((dead13:7.000000,(dead19:9.000000,(sp52:30.000000,(sp53:25.000000,sp54:25.000000)node20:5.000000)node17:5.000000)node16:11.000000)node9:5.000000,((dead32:7.000000,(sp60:14.000000,(sp70:6.000000,(sp85:1.000000,sp86:1.000000)node49:5.000000)node40:8.000000)node31:7.000000)node22:25.000000,sp50:46.000000)node8:5.000000)node6:15.000000)node3:3.000000)node1:3.000000,((((dead21:13.000000,(sp61:13.000000,sp62:13.000000)node33:23.000000)node15:4.000000,((((dead42:12.000000,(sp69:6.000000,(sp83:1.000000,sp84:1.000000)node48:5.000000)node39:10.000000)node29:3.000000,(sp81:1.000000,sp82:1.000000)node47:18.000000)node26:7.000000,(sp58:18.000000,(sp66:9.000000,(sp75:3.000000,sp76:3.000000)node44:6.000000)node37:9.000000)node27:8.000000)node18:11.000000,(sp77:1.000000,sp78:1.000000)node45:36.000000)node14:3.000000)node11:3.000000,(sp57:20.000000,(sp59:18.000000,(((sp63:11.000000,(sp64:10.000000,sp65:10.000000)node36:1.000000)node35:2.000000,(sp67:7.000000,sp68:7.000000)node38:6.000000)node34:3.000000,(sp73:4.000000,sp74:4.000000)node43:12.000000)node30:2.000000)node28:2.000000)node24:23.000000)node10:8.000000,((sp51:39.000000,((sp71:5.000000,sp72:5.000000)node41:15.000000,(sp79:1.000000,sp80:1.000000)node46:19.000000)node25:19.000000)node12:10.000000,(sp55:20.000000,sp56:20.000000)node23:29.000000)node7:2.000000)node5:21.000000)node0:11.000000;\n"
#> 
#> $sample
#> # A tibble: 37 × 3
#>    sample abundance  name
#>     <chr>     <int> <chr>
#> 1   alive         1  sp50
#> 2   alive         1  sp51
#> 3   alive         1  sp52
#> 4   alive         1  sp53
#> 5   alive         1  sp54
#> 6   alive         1  sp55
#> 7   alive         1  sp56
#> 8   alive         1  sp57
#> 9   alive         1  sp58
#> 10  alive         1  sp59
#> # ... with 27 more rows
#> 
#> $traits
#> # A tibble: 44 × 6
#>      name ecov0 ecov1 ecov2 ecov3 ecov4
#>     <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   dead2    -3    -3    -8     4     5
#> 2   dead4    -3    -1    -1    -8     5
#> 3  dead13     6     6     1     2    11
#> 4  dead19     1    14    -7     2     4
#> 5  dead21   -13   -10   -26     6    -7
#> 6  dead32    -7    21    18    -4    -1
#> 7  dead42   -18     8   -34     8    -1
#> 8    sp50     8     9    16    13     3
#> 9    sp51   -18    -4   -14     5    -4
#> 10   sp52   -10     7     8     5     8
#> # ... with 34 more rows
#> 

ph_ecovolve(extant_lineages = FALSE)

#> $phylogeny
#> [1] "(dead1:6.000000,(((dead6:9.000000,dead7:16.000000)node5:10.000000,(sp11:5.000000,sp12:5.000000)node8:31.000000)node3:1.000000,(sp9:29.000000,sp10:29.000000)node4:8.000000)node2:26.000000)node0:37.000000;\n"
#> 
#> $sample
#> # A tibble: 4 × 3
#>   sample abundance  name
#>    <chr>     <int> <chr>
#> 1  alive         1   sp9
#> 2  alive         1  sp10
#> 3  alive         1  sp11
#> 4  alive         1  sp12
#> 
#> $traits
#> # A tibble: 7 × 6
#>    name ecov0 ecov1 ecov2 ecov3 ecov4
#>   <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 dead1   -15   -13     7     2    14
#> 2 dead6   -17     5    -4    18     5
#> 3 dead7   -17     9    14     8     6
#> 4   sp9    -8    -4    12    10    22
#> 5  sp10    -2   -10    20    28    17
#> 6  sp11   -11     3     0    10     6
#> 7  sp12   -11     5    -1    14     1
#> 

ph_ecovolve(only_extant = FALSE)

#> $phylogeny
#> [1] "((dead6:9.000000,dead7:17.000000)node3:24.000000,((((dead12:5.000000,((dead23:14.000000,sp30:24.000000)node18:9.000000,(sp31:14.000000,(sp33:9.000000,sp34:9.000000)node25:5.000000)node21:19.000000)node14:7.000000)node9:5.000000,dead13:12.000000)node8:17.000000,(dead20:41.000000,dead24:49.000000)node5:4.000000)node4:18.000000,((dead15:5.000000,sp27:36.000000)node11:42.000000,((dead19:10.000000,(sp32:10.000000,(sp35:9.000000,sp36:9.000000)node26:1.000000)node22:19.000000)node16:8.000000,(sp28:27.000000,sp29:27.000000)node17:10.000000)node10:41.000000)node2:2.000000)node1:8.000000)node0:12.000000;\n"
#> 
#> $sample
#> # A tibble: 10 × 3
#>    sample abundance  name
#>     <chr>     <int> <chr>
#> 1   alive         1  sp27
#> 2   alive         1  sp28
#> 3   alive         1  sp29
#> 4   alive         1  sp30
#> 5   alive         1  sp31
#> 6   alive         1  sp32
#> 7   alive         1  sp33
#> 8   alive         1  sp34
#> 9   alive         1  sp35
#> 10  alive         1  sp36
#> 
#> $traits
#> # A tibble: 19 × 6
#>      name ecov0 ecov1 ecov2 ecov3 ecov4
#>     <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   dead6    22     0    10    -5     2
#> 2   dead7    17    -5     7     6   -11
#> 3  dead12    18   -10    26    12    -9
#> 4  dead13    22   -21    31     9   -10
#> 5  dead15    12     8    16     4   -10
#> 6  dead19    26   -12   -25    11     2
#> 7  dead20    25    -6    18    23   -18
#> 8  dead23    20    -2    26     3   -10
#> 9  dead24    32   -18    -1   -16   -12
#> 10   sp27    15     9    -1    16   -26
#> 11   sp28    37    -5   -37     5    17
#> 12   sp29    28     2   -18     1    16
#> 13   sp30    20   -18     8     2     4
#> 14   sp31    15     9    25    -7   -10
#> 15   sp32    40   -11   -39    28     9
#> 16   sp33     5    12    32    -3    -4
#> 17   sp34    21     8    19   -11    -5
#> 18   sp35    38    -9   -32    16    18
#> 19   sp36    50   -12   -32    14    22
#> 

# ph_ecovolve(only_extant = TRUE)
ph_ecovolve(taper_change = 2)

#> $phylogeny
#> [1] "(dead1:4.000000,((dead10:25.000000,(((dead14:10.000000,((sp69:10.000000,sp70:10.000000)node39:7.000000,(sp71:9.000000,sp72:9.000000)node41:8.000000)node29:25.000000)node9:1.000000,((dead32:1.000000,dead43:9.000000)node30:8.000000,sp55:24.000000)node20:19.000000)node7:6.000000,(((dead22:14.000000,(((dead48:10.000000,(sp65:11.000000,sp66:11.000000)node37:1.000000)node34:19.000000,(sp54:29.000000,(sp67:10.000000,sp68:10.000000)node38:19.000000)node18:2.000000)node16:4.000000,((sp53:32.000000,((sp56:23.000000,(((sp63:11.000000,sp64:11.000000)node36:7.000000,((sp79:4.000000,sp80:4.000000)node46:5.000000,(sp83:2.000000,sp84:2.000000)node49:7.000000)node40:9.000000)node28:3.000000,(sp77:6.000000,sp78:6.000000)node45:15.000000)node26:2.000000)node23:7.000000,((sp59:22.000000,sp60:22.000000)node25:3.000000,(sp62:13.000000,(sp75:7.000000,sp76:7.000000)node44:6.000000)node33:12.000000)node19:5.000000)node17:2.000000)node15:2.000000,(sp57:22.000000,sp58:22.000000)node24:12.000000)node13:1.000000)node12:2.000000)node11:5.000000,sp52:42.000000)node8:5.000000,(((dead35:7.000000,sp61:19.000000)node27:4.000000,(sp81:2.000000,sp82:2.000000)node47:21.000000)node21:21.000000,((sp73:8.000000,sp74:8.000000)node42:7.000000,(sp85:1.000000,sp86:1.000000)node50:14.000000)node31:29.000000)node6:3.000000)node5:2.000000)node4:13.000000)node3:14.000000,sp51:76.000000)node2:5.000000)node0:19.000000;\n"
#> 
#> $sample
#> # A tibble: 36 × 3
#>    sample abundance  name
#>     <chr>     <int> <chr>
#> 1   alive         1  sp51
#> 2   alive         1  sp52
#> 3   alive         1  sp53
#> 4   alive         1  sp54
#> 5   alive         1  sp55
#> 6   alive         1  sp56
#> 7   alive         1  sp57
#> 8   alive         1  sp58
#> 9   alive         1  sp59
#> 10  alive         1  sp60
#> # ... with 26 more rows
#> 
#> $traits
#> # A tibble: 44 × 6
#>      name ecov0 ecov1 ecov2 ecov3 ecov4
#>     <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   dead1    13     0    -9     6   -10
#> 2  dead10    25    15    14    25   -19
#> 3  dead14    38    10    -5    16   -16
#> 4  dead22    35    16    -5    21   -13
#> 5  dead32    31    14    -4     8    -3
#> 6  dead35    20    11     3    15     0
#> 7  dead43    33     2    -4     7     4
#> 8  dead48    21   -12   -14    -7    -3
#> 9    sp51    24    -3    24    16    -2
#> 10   sp52    30    28    14    18    -7
#> # ... with 34 more rows
#> 

ph_ecovolve(taper_change = 10)

#> $phylogeny
#> [1] "(sp5:14.000000,((sp6:2.000000,(sp9:1.000000,sp10:1.000000)node4:1.000000)node2:6.000000,(sp7:1.000000,sp8:1.000000)node3:7.000000)node1:6.000000)node0:86.000000;\n"
#> 
#> $sample
#> # A tibble: 6 × 3
#>   sample abundance  name
#>    <chr>     <int> <chr>
#> 1  alive         1   sp5
#> 2  alive         1   sp6
#> 3  alive         1   sp7
#> 4  alive         1   sp8
#> 5  alive         1   sp9
#> 6  alive         1  sp10
#> 
#> $traits
#> # A tibble: 6 × 6
#>    name ecov0 ecov1 ecov2 ecov3 ecov4
#>   <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   sp5    -5   -10    -7     0     3
#> 2   sp6   -18   -12   -20    -8   -13
#> 3   sp7   -17   -11   -20   -11    -9
#> 4   sp8   -17   -11   -20   -11    -9
#> 5   sp9   -16   -12   -18   -13   -14
#> 6  sp10   -16   -12   -18   -13   -14
#> 

ph_ecovolve(taper_change = 500)

#> $phylogeny
#> [1] "(dead1:7.000000,(((dead8:4.000000,sp17:12.000000)node6:1.000000,sp16:13.000000)node4:15.000000,(((sp18:9.000000,sp19:9.000000)node7:4.000000,(sp20:7.000000,(sp21:6.000000,(sp22:5.000000,(sp23:4.000000,(sp25:2.000000,sp26:2.000000)node14:2.000000)node12:1.000000)node11:1.000000)node10:1.000000)node9:6.000000)node5:6.000000,(sp24:2.000000,(sp27:1.000000,sp28:1.000000)node15:1.000000)node13:17.000000)node3:9.000000)node2:58.000000)node0:14.000000;\n"
#> 
#> $sample
#> # A tibble: 13 × 3
#>    sample abundance  name
#>     <chr>     <int> <chr>
#> 1   alive         1  sp16
#> 2   alive         1  sp17
#> 3   alive         1  sp18
#> 4   alive         1  sp19
#> 5   alive         1  sp20
#> 6   alive         1  sp21
#> 7   alive         1  sp22
#> 8   alive         1  sp23
#> 9   alive         1  sp24
#> 10  alive         1  sp25
#> 11  alive         1  sp26
#> 12  alive         1  sp27
#> 13  alive         1  sp28
#> 
#> $traits
#> # A tibble: 15 × 6
#>     name ecov0 ecov1 ecov2 ecov3 ecov4
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  dead1    -4     1     5    -1    -2
#> 2  dead8    19    -2    39    -4   -23
#> 3   sp16    19    -4    38    -8   -15
#> 4   sp17    34     1    34    -2   -19
#> 5   sp18    19     9    40   -13   -14
#> 6   sp19    15    -3    47    -8   -25
#> 7   sp20    19    -3    34   -15   -13
#> 8   sp21    23     3    40   -11    -7
#> 9   sp22    20    -2    45   -11   -15
#> 10  sp23    26    -4    49   -12    -7
#> 11  sp24    17    -3    48    -8   -12
#> 12  sp25    24    -6    51   -14    -9
#> 13  sp26    25    -5    53   -13    -6
#> 14  sp27    14    -3    50    -8   -13
#> 15  sp28    14    -3    50    -8   -13
#> 

ph_ecovolve(competition = TRUE)

#> $phylogeny
#> [1] "(dead2:4.000000,(dead3:2.000000,(((dead9:1.000000,(dead105:3.000000,(sp179:10.000000,(sp185:7.000000,sp186:7.000000)node121:3.000000)node111:5.000000)node93:55.000000)node8:16.000000,(dead12:16.000000,(dead16:10.000000,((dead25:3.000000,(dead49:4.000000,sp159:35.000000)node39:13.000000)node21:3.000000,(dead44:3.000000,sp158:36.000000)node36:15.000000)node18:15.000000)node10:13.000000)node7:7.000000)node5:6.000000,((dead14:3.000000,(((dead99:8.000000,((dead122:2.000000,(dead135:2.000000,(sp202:2.000000,sp203:2.000000)node145:4.000000)node124:3.000000)node114:4.000000,(dead126:6.000000,sp174:11.000000)node107:2.000000)node103:9.000000)node74:9.000000,sp162:31.000000)node50:1.000000,sp160:32.000000)node45:31.000000)node11:16.000000,((((dead23:5.000000,((((dead33:1.000000,((dead100:6.000000,sp166:20.000000)node83:1.000000,((dead155:1.000000,sp204:2.000000)node147:2.000000,(sp196:3.000000,sp197:3.000000)node141:1.000000)node128:17.000000)node78:18.000000)node30:3.000000,((((dead71:4.000000,dead150:26.000000)node59:3.000000,dead142:28.000000)node52:5.000000,((dead77:4.000000,(dead81:2.000000,(dead134:5.000000,sp180:9.000000)node113:14.000000)node72:3.000000)node62:6.000000,sp161:32.000000)node47:3.000000)node40:3.000000,(((dead82:4.000000,sp165:24.000000)node69:8.000000,(sp163:28.000000,((sp173:11.000000,(sp184:7.000000,(sp194:3.000000,sp195:3.000000)node140:4.000000)node120:4.000000)node106:11.000000,(sp198:2.000000,sp199:2.000000)node143:20.000000)node75:6.000000)node56:4.000000)node48:1.000000,((((dead117:5.000000,(sp181:9.000000,(sp187:7.000000,sp188:7.000000)node123:2.000000)node115:4.000000)node104:3.000000,sp170:16.000000)node92:3.000000,sp167:19.000000)node84:7.000000,(((dead119:7.000000,dead131:10.000000)node102:1.000000,(sp183:7.000000,(sp200:2.000000,sp201:2.000000)node144:5.000000)node118:8.000000)node95:4.000000,((((dead156:1.000000,(sp209:1.000000,sp210:1.000000)node157:1.000000)node149:1.000000,sp193:3.000000)node139:5.000000,sp182:8.000000)node116:7.000000,sp171:15.000000)node96:4.000000)node85:7.000000)node61:7.000000)node43:5.000000)node32:4.000000)node28:2.000000,dead51:14.000000)node26:4.000000,(((((dead60:1.000000,dead67:3.000000)node55:4.000000,dead98:18.000000)node46:7.000000,(dead73:15.000000,((((dead153:1.000000,(sp207:1.000000,sp208:1.000000)node154:1.000000)node146:3.000000,sp189:5.000000)node127:5.000000,sp175:10.000000)node108:15.000000,(sp191:3.000000,sp192:3.000000)node136:22.000000)node66:12.000000)node34:2.000000)node29:4.000000,(dead63:2.000000,dead94:13.000000)node54:15.000000)node27:3.000000,(dead64:1.000000,(dead87:3.000000,dead125:17.000000)node76:5.000000)node57:19.000000)node24:2.000000)node22:3.000000)node19:3.000000,dead42:20.000000)node17:5.000000,(dead37:13.000000,(dead130:13.000000,sp169:17.000000)node90:32.000000)node20:10.000000)node15:1.000000,(((dead41:1.000000,(dead129:17.000000,(dead137:15.000000,(((dead148:1.000000,dead152:2.000000)node138:7.000000,(sp190:4.000000,(sp205:1.000000,sp206:1.000000)node151:3.000000)node133:6.000000)node112:4.000000,sp172:14.000000)node101:4.000000)node88:3.000000)node79:15.000000)node38:1.000000,((dead80:5.000000,(sp164:25.000000,(sp176:10.000000,sp177:10.000000)node109:15.000000)node68:1.000000)node65:1.000000,(dead91:8.000000,(dead97:4.000000,((dead132:6.000000,sp178:10.000000)node110:8.000000,sp168:18.000000)node89:1.000000)node86:5.000000)node70:3.000000)node58:10.000000)node35:1.000000,dead53:9.000000)node31:22.000000)node13:19.000000)node6:13.000000)node4:4.000000)node1:2.000000)node0:2.000000;\n"
#> 
#> $sample
#> # A tibble: 53 × 3
#>    sample abundance  name
#>     <chr>     <int> <chr>
#> 1   alive         1 sp158
#> 2   alive         1 sp159
#> 3   alive         1 sp160
#> 4   alive         1 sp161
#> 5   alive         1 sp162
#> 6   alive         1 sp163
#> 7   alive         1 sp164
#> 8   alive         1 sp165
#> 9   alive         1 sp166
#> 10  alive         1 sp167
#> # ... with 43 more rows
#> 
#> $traits
#> # A tibble: 106 × 6
#>      name ecov0 ecov1 ecov2 ecov3 ecov4
#>     <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   dead2    -3    -7    -4     1     4
#> 2   dead3    -6    -3     3     7     4
#> 3   dead9     2   -15    -1    -8    -3
#> 4  dead12    -7   -31     4    -5     5
#> 5  dead14    -2   -13    -2   -11    -3
#> 6  dead16     2   -27     5     7   -14
#> 7  dead23   -13   -30    -2     0     5
#> 8  dead25    -9   -21    -3     7     1
#> 9  dead33   -12   -25     0     3     3
#> 10 dead37    -6   -10     1    -5     6
#> # ... with 96 more rows
#> 

ph_ecovolve(competition = FALSE)

#> $phylogeny
#> [1] "(((dead7:9.000000,(((((dead52:3.000000,(sp110:8.000000,(sp120:7.000000,(sp125:5.000000,sp126:5.000000)node70:2.000000)node66:1.000000)node59:6.000000)node45:9.000000,((dead74:9.000000,sp99:12.000000)node47:7.000000,sp94:19.000000)node40:4.000000)node37:9.000000,(sp113:7.000000,sp114:7.000000)node62:25.000000)node23:24.000000,(dead77:17.000000,(sp108:8.000000,sp109:8.000000)node58:11.000000)node39:37.000000)node8:6.000000,(((dead85:7.000000,sp111:8.000000)node60:1.000000,(sp145:1.000000,sp146:1.000000)node83:8.000000)node55:4.000000,sp98:13.000000)node46:49.000000)node4:3.000000)node2:4.000000,(((dead29:6.000000,(sp95:18.000000,sp96:18.000000)node41:16.000000)node21:12.000000,(((sp102:10.000000,sp103:10.000000)node53:24.000000,((sp117:7.000000,sp118:7.000000)node64:19.000000,(sp128:3.000000,sp129:3.000000)node72:23.000000)node33:8.000000)node22:10.000000,(sp105:8.000000,(sp147:1.000000,sp148:1.000000)node84:7.000000)node56:36.000000)node16:2.000000)node15:14.000000,(sp89:38.000000,(sp97:17.000000,(sp100:12.000000,(sp104:10.000000,(sp137:2.000000,sp138:2.000000)node79:8.000000)node54:2.000000)node48:5.000000)node42:21.000000)node18:22.000000)node5:9.000000)node1:16.000000,((dead9:3.000000,((((dead34:1.000000,sp93:26.000000)node31:25.000000,(((sp91:29.000000,(sp127:3.000000,(sp149:1.000000,sp150:1.000000)node86:2.000000)node71:26.000000)node27:9.000000,(sp112:7.000000,(sp132:3.000000,(sp151:1.000000,sp152:1.000000)node87:2.000000)node75:4.000000)node61:31.000000)node20:10.000000,(sp139:1.000000,sp140:1.000000)node80:47.000000)node14:3.000000)node13:1.000000,sp88:52.000000)node12:3.000000,((dead38:2.000000,(sp133:2.000000,sp134:2.000000)node76:22.000000)node36:4.000000,dead50:17.000000)node28:27.000000)node10:3.000000)node6:5.000000,((((dead43:15.000000,(sp92:27.000000,(sp115:7.000000,sp116:7.000000)node63:20.000000)node30:3.000000)node25:1.000000,(sp106:8.000000,sp107:8.000000)node57:23.000000)node24:7.000000,(((((dead69:2.000000,sp119:7.000000)node65:5.000000,(sp135:2.000000,sp136:2.000000)node78:10.000000)node49:3.000000,(sp141:1.000000,sp142:1.000000)node81:14.000000)node44:9.000000,(sp130:3.000000,sp131:3.000000)node73:21.000000)node35:5.000000,sp90:29.000000)node26:9.000000)node19:14.000000,(((sp101:11.000000,(sp143:1.000000,sp144:1.000000)node82:10.000000)node51:15.000000,(sp121:6.000000,sp122:6.000000)node67:20.000000)node32:15.000000,(sp123:5.000000,sp124:5.000000)node68:36.000000)node17:11.000000)node11:11.000000)node3:22.000000)node0:15.000000;\n"
#> 
#> $sample
#> # A tibble: 65 × 3
#>    sample abundance  name
#>     <chr>     <int> <chr>
#> 1   alive         1  sp88
#> 2   alive         1  sp89
#> 3   alive         1  sp90
#> 4   alive         1  sp91
#> 5   alive         1  sp92
#> 6   alive         1  sp93
#> 7   alive         1  sp94
#> 8   alive         1  sp95
#> 9   alive         1  sp96
#> 10  alive         1  sp97
#> # ... with 55 more rows
#> 
#> $traits
#> # A tibble: 77 × 6
#>      name ecov0 ecov1 ecov2 ecov3 ecov4
#>     <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   dead7    15   -21   -20     8    -1
#> 2   dead9    11    -3   -29     4    20
#> 3  dead29     6   -12   -12     6     6
#> 4  dead34    -1     3   -40     0    24
#> 5  dead38    27     8   -34     6     3
#> 6  dead43     7   -12   -37   -10    -3
#> 7  dead50    36    14   -23     6     2
#> 8  dead52    14     4   -11    -5    -5
#> 9  dead69    27     0   -44   -13     9
#> 10 dead74     1     5   -22     2     2
#> # ... with 67 more rows
#> 

# library(ape)
# x <- ph_ecovolve(speciation = 0.05)
# plot(read.tree(text = x))