rao - Rao's quadratic entropy

A measure of within- and among-community diversity taking species dissimilarity (phylogenetic dissimilarity) into account

ph_rao(sample, phylo)

Arguments

sample

(data.frame/character) sample data.frame or path to a sample file

phylo

(character/phylo) phylogeny as a phylo object or a newick string (will be written to a temp file if provided) - or a path to a file with a newick string

Value

A list of 6 data.frame's: Diversity components:

overall alpha (within-site)

beta (among-site)

total diversity

Fst statistic of differentiation for diversity and phylogenetic diversity

Within-community diversity:

Plot - Plot name

NSpp - Number of species

NIndiv - Number of individuals

PropIndiv - Proportion of all individuals found in this plot

D - Diversity (= Simpson’s diversity)

Dp - Phylogenetic diversity (= Diversity weighted by interspecific phylogenetic distances)

The remaining 4 tables compare each community pairwise:

among_community_diversity_d - Among-community diversities

among_community_diversity_h - Among-community diversities excluding within-community diversity

among_community_phylogenetic_diversity_dp - Among-community phylogenetic diversities

among_community_phylogenetic_diversity_hp - Among-community phylogenetic diversities excluding within-community diversity

See also

Other phylogenetic-diversity: ph_pd

Examples

sfile <- system.file("examples/sample_comstruct", package = "phylocomr")
pfile <- system.file("examples/phylo_comstruct", package = "phylocomr")

# sample from data.frame, phylogeny from a string
sampledf <- read.table(sfile, header = FALSE,
  stringsAsFactors = FALSE)
phylo_str <- readLines(pfile)

ph_rao(sample = sampledf, phylo = phylo_str)

#> $diversity_components
#> # A tibble: 4 × 3
#>   Component Standard Phylogenetic
#>       <chr>    <dbl>        <dbl>
#> 1     Alpha 0.860294     3.106005
#> 2      Beta 0.084775     0.752559
#> 3     Total 0.945069     3.858564
#> 4       Fst 0.089703     0.195036
#> 
#> $within_community_diveristy
#> # A tibble: 6 × 6
#>      Plot  NSpp NIndiv PropIndiv        D       Dp
#>     <chr> <int>  <int>     <dbl>    <dbl>    <dbl>
#> 1  clump1     8      8  0.117647 0.875000 2.125000
#> 2 clump2a     8     12  0.176471 0.861111 2.472222
#> 3 clump2b     8     12  0.176471 0.861111 2.916667
#> 4  clump4     8     12  0.176471 0.861111 3.472222
#> 5    even     8      8  0.117647 0.875000 3.875000
#> 6  random     8     16  0.235294 0.843750 3.554688
#> 
#> $among_community_diversity_d
#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 0.875000 0.958333 0.958333 0.979167 0.968750 0.976562
#> 2 clump2a 0.958333 0.861111 0.972222 0.958333 0.968750 0.984375
#> 3 clump2b 0.958333 0.972222 0.861111 0.930556 0.968750 0.963542
#> 4  clump4 0.979167 0.958333 0.930556 0.861111 0.937500 0.963542
#> 5    even 0.968750 0.968750 0.968750 0.937500 0.875000 0.960938
#> 6  random 0.976562 0.984375 0.963542 0.963542 0.960938 0.843750
#> 
#> $among_community_diversity_h
#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 0.000000 0.090278 0.090278 0.111111 0.093750 0.117188
#> 2 clump2a 0.090278 0.000000 0.111111 0.097222 0.100694 0.131944
#> 3 clump2b 0.090278 0.111111 0.000000 0.069444 0.100694 0.111111
#> 4  clump4 0.111111 0.097222 0.069444 0.000000 0.069444 0.111111
#> 5    even 0.093750 0.100694 0.100694 0.069444 0.000000 0.101562
#> 6  random 0.117188 0.131944 0.111111 0.111111 0.101562 0.000000
#> 
#> $among_community_phylogenetic_diversity_dp
#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4    even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>
#> 1  clump1 2.125000 3.375000 4.041667 4.354167 4.03125 4.023438
#> 2 clump2a 3.375000 2.472222 4.361111 4.208334 4.03125 3.911458
#> 3 clump2b 4.041667 4.361111 2.916667 3.680556 4.03125 3.973958
#> 4  clump4 4.354167 4.208334 3.680556 3.472222 3.93750 4.119792
#> 5    even 4.031250 4.031250 4.031250 3.937500 3.87500 4.000000
#> 6  random 4.023438 3.911458 3.973958 4.119792 4.00000 3.554688
#> 
#> $among_community_phylogenetic_diversity_hp
#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 0.000000 1.076389 1.520833 1.555556 1.031250 1.183594
#> 2 clump2a 1.076389 0.000000 1.666667 1.236111 0.857639 0.898004
#> 3 clump2b 1.520833 1.666667 0.000000 0.486111 0.635417 0.738281
#> 4  clump4 1.555556 1.236111 0.486111 0.000000 0.263889 0.606337
#> 5    even 1.031250 0.857639 0.635417 0.263889 0.000000 0.285156
#> 6  random 1.183594 0.898004 0.738281 0.606337 0.285156 0.000000
#> 

# both from files
sample_str <- paste0(readLines(sfile), collapse = "\n")
sfile2 <- tempfile()
cat(sample_str, file = sfile2, sep = '\n')
pfile2 <- tempfile()
phylo_str <- readLines(pfile)
cat(phylo_str, file = pfile2, sep = '\n')

ph_rao(sample = sfile2, phylo = pfile2)

#> $diversity_components
#> # A tibble: 4 × 3
#>   Component Standard Phylogenetic
#>       <chr>    <dbl>        <dbl>
#> 1     Alpha 0.860294     3.106005
#> 2      Beta 0.084775     0.752559
#> 3     Total 0.945069     3.858564
#> 4       Fst 0.089703     0.195036
#> 
#> $within_community_diveristy
#> # A tibble: 6 × 6
#>      Plot  NSpp NIndiv PropIndiv        D       Dp
#>     <chr> <int>  <int>     <dbl>    <dbl>    <dbl>
#> 1  clump1     8      8  0.117647 0.875000 2.125000
#> 2 clump2a     8     12  0.176471 0.861111 2.472222
#> 3 clump2b     8     12  0.176471 0.861111 2.916667
#> 4  clump4     8     12  0.176471 0.861111 3.472222
#> 5    even     8      8  0.117647 0.875000 3.875000
#> 6  random     8     16  0.235294 0.843750 3.554688
#> 
#> $among_community_diversity_d
#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 0.875000 0.958333 0.958333 0.979167 0.968750 0.976562
#> 2 clump2a 0.958333 0.861111 0.972222 0.958333 0.968750 0.984375
#> 3 clump2b 0.958333 0.972222 0.861111 0.930556 0.968750 0.963542
#> 4  clump4 0.979167 0.958333 0.930556 0.861111 0.937500 0.963542
#> 5    even 0.968750 0.968750 0.968750 0.937500 0.875000 0.960938
#> 6  random 0.976562 0.984375 0.963542 0.963542 0.960938 0.843750
#> 
#> $among_community_diversity_h
#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 0.000000 0.090278 0.090278 0.111111 0.093750 0.117188
#> 2 clump2a 0.090278 0.000000 0.111111 0.097222 0.100694 0.131944
#> 3 clump2b 0.090278 0.111111 0.000000 0.069444 0.100694 0.111111
#> 4  clump4 0.111111 0.097222 0.069444 0.000000 0.069444 0.111111
#> 5    even 0.093750 0.100694 0.100694 0.069444 0.000000 0.101562
#> 6  random 0.117188 0.131944 0.111111 0.111111 0.101562 0.000000
#> 
#> $among_community_phylogenetic_diversity_dp
#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4    even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>
#> 1  clump1 2.125000 3.375000 4.041667 4.354167 4.03125 4.023438
#> 2 clump2a 3.375000 2.472222 4.361111 4.208334 4.03125 3.911458
#> 3 clump2b 4.041667 4.361111 2.916667 3.680556 4.03125 3.973958
#> 4  clump4 4.354167 4.208334 3.680556 3.472222 3.93750 4.119792
#> 5    even 4.031250 4.031250 4.031250 3.937500 3.87500 4.000000
#> 6  random 4.023438 3.911458 3.973958 4.119792 4.00000 3.554688
#> 
#> $among_community_phylogenetic_diversity_hp
#> # A tibble: 6 × 7
#>         .   clump1  clump2a  clump2b   clump4     even   random
#>     <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1  clump1 0.000000 1.076389 1.520833 1.555556 1.031250 1.183594
#> 2 clump2a 1.076389 0.000000 1.666667 1.236111 0.857639 0.898004
#> 3 clump2b 1.520833 1.666667 0.000000 0.486111 0.635417 0.738281
#> 4  clump4 1.555556 1.236111 0.486111 0.000000 0.263889 0.606337
#> 5    even 1.031250 0.857639 0.635417 0.263889 0.000000 0.285156
#> 6  random 1.183594 0.898004 0.738281 0.606337 0.285156 0.000000
#>